fix: Remove type cast 'as int' from String UUID in dashboard screen

+++ community_hero_app/lib/features/dashboard/presentation/screens/dashboard_screen.dart
@@ -610,7 +610,7 @@ class _RecentActivityList extends ConsumerWidget {
                         onTap: () {
                           final id = item['id'];
                           if (id != null) {
-                            ref.read(dashboardControllerProvider.notifier).resolveIssue(id as int);
+                            ref.read(dashboardControllerProvider.notifier).resolveIssue(id);
                           }
                         },
                         borderRadius: BorderRadius.circular(12),
